implement 3d-histogram; implement widget to expand; refectoring; optimize performance; implement normalize 3D-hisotgram; elimination of various errors; implement recursiv lambda to find some controlwidgets; better implementation of YUVFactory

+++ source/chartHandlerTypedef.h
@@ -327,7 +327,7 @@ struct chartSettingsData
   QAbstractSeries* mSeries;
 
   //for 3D data we use
-  QMap<int, QMap<int, int>> m3DData;
+  QMap<int, QMap<int, double>> m3DData;
 };
